start-arm: disable I-cache
<HACK>

For some yet unknown reason the processor on the i.MX35 3stack board (at
least on our board) has problem with an activated instruction cache and
booting from NAND: The copy-from-NFC-RAM-to-SDRAM routine doesn't loop.

It looks basically like this:

1:
	ldmia
	stmia
	cmp
	ble 1b

If the "cmp" instruction lives on address 0xbb000640 it doesn't work
with an activated i-cache. The processor flags are not properly updated.
So the ble won't jump back. This obviously break booting from nand.

This is why this patch disables the i-cache.

</HACK>

Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>