fb: sdl: create graphics window on enabling
The framebuffer should be enabled on the fb_enable callback and disabled
in the fb_disable callback. In SDL context this means the grapics window
has to be created / destroyed in fb_enable / fb_disable.

With this change the framebuffer has to be enabled explicitly with
fb0.enable=1 like with other framebuffer driver aswell.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>