partitions: UUID was missing from partitions by name
The partition code registers one or two cdevs for each partition.  The
2nd one is used if the partition has a name.  The block of code for
each cdev is duplicated twice, and the 2nd copy is missing the bit
that copies the partition uuid and the dos partition type to the cdev.

Refactor the register cdev out to a new function that is called twice.
This way each cdev is done correctly.

Signed-off-by: Trent Piepho <tpiepho@kymetacorp.com>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>