base: add class device support
This introduces the concept of class devices in barebox. Class devices
group together devices of the same type. Several subsystems like
watchdog and network devices have a struct device embedded in their
subsystem specific struct anyway, so we can use this as a class device.
As these class devices are collected on a list we can use this list to
iterate over all network/watchdog devices and thus free the subsystems
from the burden of keeping a list themselves.

There is a 'class' command added in this patch which can be used to show
all registered classes along with the devices registered for each class.

The classes known at compile time can be collected on a linker list
using the macro DEFINE_DEV_CLASS(). That way registering classes does
not need an additional initcall and also no care has to be taken that
this initcalls runs before the first user adds devices to a new class.

Link: https://lore.barebox.org/20240611051237.4126918-3-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>